% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testenv.R
\name{simplefun}
\alias{simplefun}
\title{testing role of environment variables in check surveillance}
\usage{
simplefun(x)
}
\arguments{
\item{x}{numeric()}
}
\description{
testing role of environment variables in check surveillance
}
\examples{
x = rnorm(20)
simplefun(x)
}
